+++ app/assets/javascripts/blob/csv_viewer.js
@@ -1,0 +1,3 @@
+import renderCSV from './csv';
+
+export default renderCSV;